redesign faq section contact us and folder structure rebuild
