+++ src/app/layout.js
@@ -5,11 +5,8 @@ import './globals.css'
 import favicon from '../../public/images/favicon.png'
 import Head from 'next/head';
 import './globals.css';
-import Navbar from '@/components/Home/Navbar';
-import Footer from '@/components/Home/Footer';
-// import Navbar from '@/components/home/Navbar';
-// import Footer from '@/components/home/Footer';
-// import { NextUIProvider } from '@nextui-org/react';
+import Footer from '@/shared/Footer';
+import Navbar from '@/shared/Navbar';
 
 
 const inter = Inter({ subsets: ['latin'] });
@@ -28,9 +25,7 @@ const RootLayout = ({ children }) => {
         <link rel="shortcut icon" href={favicon} />
         <title>{metadata.title}</title>
       </Head>
-
       <body className={inter.className}>
-
         <Navbar />
         {children}
         <Footer />
