+++ src/pages/about/AboutMission.jsx
