+++ src/pages/home/Slider/Slider.js
@@ -3,7 +3,7 @@ import React from "react";
 import Swiper from "./Swiper";
 const Slider = () => {
   return (
-    <section className="container mx-auto my-16 ">
+    <section className="container mx-auto mt-16">
       <p className="text-3xl font-bold text-center mb-10">Our Top Author &apos;s</p>
       <Swiper></Swiper>
     </section>
