+++ src/pages/home/faq/faq.css
@@ -1,0 +1,25 @@
+.text-foreground {
+    text-align: left;
+    font-weight: 600;
+    margin: 10px 0;
+}
+
+@media (min-width: 768px) {
+    .text-foreground {
+        font-size: 1rem;
+    }
+}
+
+
+
+
+@media (max-width: 500px) {
+    .text-foreground {
+        font-size: .9rem;
+        line-height: 1.2;
+    }
+
+    .faq {
+        font-size: .9rem;
+    }
+}
